requirements.txt
langchain
ctransformers
sentence-transformers
faiss-cpu
llama-cpp-python
googletrans==4.0.0rc1
Django
openai-whisper